==> data-prep/1. Format Data for Recipe Database.ipynb <==
import pickle
import pandas as pd
from ast import literal_eval

## Ingredient Files
Convert ingr_map.pkl into cleanIngredients.csv and rawIngredients.csv

# Import ingredient data
ingr_map_file = open('./Recipe Data/ingr_map.pkl','rb')
ingr_map_dict = pickle.load(ingr_map_file)
ingr_map_file.close()

## convert to dataframe
ingr_map_df = pd.DataFrame(ingr_map_dict)
ingr_map_df.head()

### cleanIngredients.csv

# remove duplicates
cleanIngredients = ingr_map_df.loc[:,["replaced", "id"]].copy().drop_duplicates()

# rename columns
cleanIngredients.rename({"replaced": "name", "id": "cleanIngredientID"}, axis=1, inplace=True)

# reset index
cleanIngredients.reset_index(drop=True, inplace=True)

# reorder columns
cleanIngredients = cleanIngredients[["cleanIngredientID", "name"]]

cleanIngredients.head()

# Confirm that all remaining ingredients are unique
assert len(ingr_map_df["replaced"].unique()) == cleanIngredients.shape[0]
assert len(ingr_map_df["id"].unique()) == cleanIngredients.shape[0]

### rawIngredients.csv

rawIngredients = ingr_map_df.copy()

# drop clean ingredient information
rawIngredients.drop(["replaced", "count"], axis=1, inplace=True)

# create ID from index
rawIngredients.reset_index(inplace=True)

# rename columns
rawIngredients.rename({"index": "rawIngredientID", "raw_ingr": "rawName", "raw_words": "rawNameLength", \
                       "processed": "processedName", "len_proc": "processedNameLength", \
                       "id":"cleanIngredientID"}, axis=1, inplace=True)

rawIngredients.head()

## Recipe Files - pp_recipes.csv
Format data and create join table for recipe and ingredients.

pp_recipes = pd.read_csv("./Recipe Data/pp_recipes.csv")
pp_recipes.head()

recipes1 = pp_recipes.copy()

# Rename columns
recipes1.rename({"id": "recipeID", "name_tokens": "nameTokenList", "ingredient_tokens": "ingredientTokenList", \
               "steps_tokens": "stepTokenList", "techniques": "techniqueList", "calorie_level": "calorieLevel",
               "ingredient_ids": "cleanIngredientIDList"}, axis=1, inplace=True)
recipes1.head()

### recipeCleanJoin.csv

#extract just the necessary columns
recipeIngredientJoin = recipes1.loc[:,["recipeID", "cleanIngredientIDList"]].copy()
recipeIngredientJoin.set_index('recipeID', inplace=True)

# convert the ingredient list into separate rows
recipeIngredientJoin = pd.melt(recipeIngredientJoin.cleanIngredientIDList.apply(lambda x: pd.Series(literal_eval(x))).reset_index(), 
             id_vars=['recipeID'],
             value_name='cleanIngredientID') \
         .drop('variable', axis=1) \
         .dropna() 
recipeIngredientJoin["cleanIngredientID"] = recipeIngredientJoin["cleanIngredientID"].astype(int)

# # create an index for the new table
# recipeIngredientJoin.reset_index(inplace=True)

# # confirm that the index is unique
# assert len(recipeIngredientJoin["index"].unique()) == recipeIngredientJoin.shape[0]

# # rename the index
# recipeIngredientJoin.rename({"index": "recipeIngredientID"}, axis=1, inplace=True)

recipeIngredientJoin.head()

## Recipe Files - raw_recipes.csv
Format data, create join table for tags, and combine with pp_recipes.csv file

raw_recipes = pd.read_csv("./Recipe Data/RAW_recipes.csv")
raw_recipes.head()

recipes2 = raw_recipes.copy()

# rename columns
recipes2.rename({"id": "recipeID", "contributor_id": "contributorID", "submitted": "submittedDate", 
                    "n_steps": "numSteps", "n_ingredients": "numIngredients", 
                 "ingredients": "cleanIngredientNameList", "nutrition": "nutritionList", "tags": "tagsList",
                "steps": "stepsList"}, 
                axis=1, inplace=True)

recipes2.head()

### recipes.csv

# confirm there are no duplicate recipe IDs in either table
assert len(recipes1["recipeID"].unique()) == recipes1.shape[0]
assert len(recipes2["recipeID"].unique()) == recipes2.shape[0]

# check out number of recipes in each
print("Records in raw_recipes.csv: ", recipes2.shape[0])
print("Records in pp_recipes.csv: ", recipes1.shape[0])

# join two recipe tables together
recipes = recipes2.merge(recipes1, how="left", on="recipeID")

# create flag to signal whether recipe data is complete or not
recipes["isComplete"] = ~pd.isnull(recipes["cleanIngredientIDList"])

# separate out nutrition data 
recipes[['calories','fatPDV', 'sugarPDV', 'sodiumPDV', 'proteinPDV', 'saturatedFatPDV', 'carbsPDV']] = \
    recipes.nutritionList.apply(lambda x: pd.Series(literal_eval(x)))

recipes.head()

### tags.csv

# extract just necessary columns
tag_data = recipes.loc[:,["recipeID", "tagsList"]].copy()
tag_data.set_index('recipeID', inplace=True)

# separate list of tags into rows
all_recipe_tags = pd.melt(tag_data.tagsList.apply(lambda x: pd.Series(literal_eval(x))).reset_index(), 
             id_vars=['recipeID'],
             value_name='tag') \
         .set_index(['recipeID']) \
         .drop('variable', axis=1) \
         .dropna() \
         .sort_index() 

all_recipe_tags.sample(5)

# extract just the necessary columns
tag_data = all_recipe_tags.copy()
tag_data["numRecipes"] = 1

# get unique list of tags and their count of recipes
tags = pd.DataFrame(tag_data.groupby("tag", as_index=False)["numRecipes"].sum())

# remove blank tags
tags = tags[tags["tag"].str.len() > 0]

# get unique index values for each tag
tags.reset_index(inplace = True)
tags.rename({"index": "tagID"}, axis=1, inplace=True)

tags

### recipeTagJoin.csv

# extract just necessary columns
all_recipe_tag_data = all_recipe_tags.copy()
all_recipe_tag_data.reset_index(inplace=True)

# join in the unique tag id
recipeTagJoin = all_recipe_tag_data.merge(tags.loc[:,["tag", "tagID"]], how="inner", on="tag")

# remove tag name
recipeTagJoin.drop("tag", inplace=True, axis=1)

# # create unique ID for table
# recipeTagJoin.reset_index(inplace=True)
# recipeTagJoin.rename({"index": "recipeTagID"}, axis=1, inplace=True)

# # confirm the IDs are unique
# assert len(recipeTagJoin["recipeTagID"].unique()) == recipeTagJoin.shape[0]

recipeTagJoin.head()

## Interaction Files

RAW_interactions = pd.read_csv("./Recipe Data/RAW_interactions.csv")
RAW_interactions.head()

### interactions.csv

interactions = RAW_interactions.copy()

# create unique ID for each interaction
interactions.reset_index(inplace=True)

# rename the columns
interactions.rename({"user_id": "userID", "recipe_id": "recipeID", "index": "interactionID"}, axis=1, inplace=True)

interactions.head()

## User Files

PP_users = pd.read_csv("./Recipe Data/PP_users.csv")
PP_users.head()

### users.csv

users = PP_users.copy()

# rename the columns
users.rename({"u": "userID", "n_items": "numItems", "n_ratings": "numRatings", "items": "itemsList", \
             "ratings": "ratingsList", "techniques": "techniquesList"}, axis=1, inplace=True)

users.head()

## Save the final files

# Save files to csv
cleanIngredients.to_csv("Clean Recipe Data/cleanIngredients.csv", index=False)
rawIngredients.to_csv("Clean Recipe Data/rawIngredients.csv", index=False)
recipeIngredientJoin.to_csv("Clean Recipe Data/recipeCleanJoin.csv", index=False)
recipes.to_csv("Clean Recipe Data/recipes.csv", index=False)
recipeTagJoin.to_csv("Clean Recipe Data/recipeTagJoin.csv", index=False)
tags.to_csv("Clean Recipe Data/tags.csv", index=False)
interactions.to_csv("Clean Recipe Data/interactions.csv", index=False)
users.to_csv("Clean Recipe Data/users.csv", index=False)
