==> data-prep/2. Update Ingredients with Groupings.ipynb <==
import pandas as pd

## Create New Ingredients Table

# Read data
cleanIngredients = pd.read_csv("Clean Recipe Data/cleanIngredients_v2.csv")
cleanIngredients.head()

# Create a new table of grouped ingredients
groupedIngredients = cleanIngredients['groupings'].unique()
groupedIngredients = pd.DataFrame(groupedIngredients, columns=['name'])

# Assign an ID
groupedIngredients.reset_index(inplace=True)
groupedIngredients.rename({'index': 'ingredientID'}, axis=1, inplace=True)

groupedIngredients.head()

## Create New Join Table

# Add new id back into old cleanIngredient table
newIngredientTable = cleanIngredients.merge(groupedIngredients, how='left', left_on='groupings', \
                                            right_on='name')
newIngredientTable = newIngredientTable.loc[:, ['cleanIngredientID', 'ingredientID']]
newIngredientTable.head()

# Import old join table
recipeCleanJoin = pd.read_csv("Clean Recipe Data/recipeCleanJoin.csv")
recipeCleanJoin.head()

# Replace old IDs with new grouped ones
recipeIngredientJoin = recipeCleanJoin.merge(newIngredientTable, how='left', \
                                             on='cleanIngredientID')
recipeIngredientJoin = recipeIngredientJoin.copy().loc[:,['recipeID', 'ingredientID']]
recipeIngredientJoin.drop_duplicates(inplace=True)
recipeIngredientJoin.head()

## Add Column with New Ingredient IDs to Recipe Table

recipeIngredientList = recipeIngredientJoin.groupby('recipeID')['ingredientID'].apply(list).reset_index(name='ingredientIDList')
recipeIngredientList.head()

## Save Files

groupedIngredients.to_csv("Clean Recipe Data/ingredients.csv", index=False)
recipeIngredientJoin.to_csv("Clean Recipe Data/recipeIngredientJoin.csv", index=False)
recipeIngredientList.to_csv("Clean Recipe Data/RecipeIngredientList.csv", index=False)
